requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
psycopg2>=2.8,<3.0
redis>=3.5,<4.0
celery>=5.0,<6.0